For consistency with other message bus related stuff, move queue related constants into transport package.
